created vacation token system

+++ assets/templates/partials_templ.go
@@ -148,7 +148,7 @@ func Event(event schemas.Event, user repo.User) templ.Component {
 	})
 }
 
-func CreateEventUpdate(event schemas.Event, user repo.User, vacationUsed int, pendingEvents int, notificationCount int) templ.Component {
+func CreateEventUpdate(event schemas.Event, user repo.User, vacationUsed float64, pendingEvents int, notificationCount int) templ.Component {
 	return templruntime.GeneratedTemplate(func(templ_7745c5c3_Input templruntime.GeneratedComponentInput) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_W, ctx := templ_7745c5c3_Input.Writer, templ_7745c5c3_Input.Context
 		if templ_7745c5c3_CtxErr := ctx.Err(); templ_7745c5c3_CtxErr != nil {
